Create Scale module
Contains: ScaleDomain, ScaleRange, ScaleNice

As this can not contain ScaleProperty (with the current module
setup), it is not clear if this is worth it.
