+++ hvega/src/Graphics/Vega/VegaLite.hs
@@ -612,6 +612,7 @@ import qualified Graphics.Vega.VegaLite.Input as VL
 import qualified Graphics.Vega.VegaLite.Legend as VL
 import qualified Graphics.Vega.VegaLite.Mark as VL
 import qualified Graphics.Vega.VegaLite.Output as VL
+import qualified Graphics.Vega.VegaLite.Scale as VL
 import qualified Graphics.Vega.VegaLite.Selection as VL
 import qualified Graphics.Vega.VegaLite.Specification as VL
 import qualified Graphics.Vega.VegaLite.Time as VL
